Move ModelException from ltr.util to (newly started) ltr.model package.
Also make tests independent of exceptions' package.
