+++ solr/contrib/ltr/src/java/org/apache/solr/ltr/ranking/LambdaMARTModel.java
@@ -25,8 +25,8 @@ import org.apache.lucene.index.LeafReaderContext;
 import org.apache.lucene.search.Explanation;
 import org.apache.solr.ltr.feature.LTRScoringModel;
 import org.apache.solr.ltr.feature.norm.Normalizer;
+import org.apache.solr.ltr.model.ModelException;
 import org.apache.solr.ltr.util.LTRUtils;
-import org.apache.solr.ltr.util.ModelException;
 
 public class LambdaMARTModel extends LTRScoringModel {
