+++ solr/contrib/ltr/src/java/org/apache/solr/ltr/rest/ManagedModelStore.java
@@ -35,9 +35,9 @@ import org.apache.solr.ltr.feature.LTRScoringModel;
 import org.apache.solr.ltr.feature.ModelStore;
 import org.apache.solr.ltr.feature.norm.Normalizer;
 import org.apache.solr.ltr.feature.norm.impl.IdentityNormalizer;
+import org.apache.solr.ltr.model.ModelException;
 import org.apache.solr.ltr.ranking.Feature;
 import org.apache.solr.ltr.util.CommonLTRParams;
-import org.apache.solr.ltr.util.ModelException;
 import org.apache.solr.response.SolrQueryResponse;
 import org.apache.solr.rest.BaseSolrResource;
 import org.apache.solr.rest.ManagedResource;
