+++ solr/contrib/ltr/src/test/org/apache/solr/ltr/TestRerankBase.java
@@ -42,12 +42,12 @@ import org.apache.solr.ltr.feature.FeatureException;
 import org.apache.solr.ltr.feature.LTRScoringModel;
 import org.apache.solr.ltr.feature.impl.ValueFeature;
 import org.apache.solr.ltr.feature.impl.ValueFeature.ValueFeatureWeight;
+import org.apache.solr.ltr.model.ModelException;
 import org.apache.solr.ltr.ranking.Feature;
 import org.apache.solr.ltr.ranking.RankSVMModel;
 import org.apache.solr.ltr.rest.ManagedFeatureStore;
 import org.apache.solr.ltr.rest.ManagedModelStore;
 import org.apache.solr.ltr.util.CommonLTRParams;
-import org.apache.solr.ltr.util.ModelException;
 import org.apache.solr.request.SolrQueryRequestBase;
 import org.apache.solr.response.SolrQueryResponse;
 import org.apache.solr.util.RestTestBase;
